backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.1
openai==1.12.0
reportlab==4.1.0
pydantic==2.6.1
pydantic-settings==2.2.1
pypdf==4.1.0
python-docx==1.1.0
razorpay==1.4.2
cryptography==42.0.2
aiofiles==23.2.1
tenacity==8.2.3
slowapi==0.1.9
email-validator==2.1.1